stub should also use typescript! Fixes some errors in stub/index.js
