+++ .gitignore
@@ -11,3 +11,6 @@ node_modules/
 yarn.lock
 npm-debug.log
 yarn-error.log
+
+# build
+dist
