+++ package.json
@@ -11,7 +11,7 @@
     "build": "webpack --mode=production",
     "test": "jest",
     "eslint": "eslint src --quiet",
-    "stub": "nodemon stub --exec babel-node --presets=@babel/preset-env"
+    "stub": "TS_NODE_COMPILER_OPTIONS='{\"module\":\"commonjs\"}' nodemon --watch 'src/**/*.ts' --exec 'ts-node' stub/index.ts"
   },
   "dependencies": {
     "@babel/runtime": "^7.9.2",
@@ -37,14 +37,15 @@
     "@babel/preset-react": "^7.9.4",
     "@babel/preset-typescript": "^7.9.0",
     "@hapi/joi": "^17.1.1",
+    "@types/hapi": "^18.0.3",
     "@types/jest": "^25.1.4",
-    "@types/lodash": "^4.14.149",
+    "@types/lodash": "^4.14.150",
     "@types/react": "^16.9.26",
     "@types/react-dom": "^16.9.5",
     "@types/react-router-dom": "^5.1.3",
     "@types/styled-components": "^5.0.1",
-    "@typescript-eslint/eslint-plugin": "^2.25.0",
-    "@typescript-eslint/parser": "^2.25.0",
+    "@typescript-eslint/eslint-plugin": "^2.29.0",
+    "@typescript-eslint/parser": "^2.29.0",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^25.2.3",
@@ -80,6 +81,7 @@
     "sinon": "^9.0.1",
     "style-loader": "^1.1.3",
     "ts-jest": "^25.2.1",
+    "ts-node": "^8.9.0",
     "webpack": "^4.42.1",
     "webpack-cli": "^3.3.11",
     "webpack-dev-server": "^3.10.3"
