+++ stub/model.ts
