+++ stub/routes.ts
@@ -1,6 +1,7 @@
+import { ServerRoute } from 'hapi'
 import { getDemo } from './model'
 
-const loadDemo = {
+const loadDemo: ServerRoute = {
   method: 'GET',
   path: '/projects',
   handler: getDemo,
