+++ tsconfig.json
@@ -24,7 +24,8 @@
     },
   },
   "include": [
-    "src/**/*"
+    "src/**/*",
+    "stub/**/*"
   ],
   "exclude": [
     "node_modules"
